+++ CHANGELOG.md
@@ -78,6 +78,9 @@ prompt is displayed.
       the Enhancements section below for details).
     - Removed `Cmd.undoc_header` since all commands are now considered categorized.
     - Renamed `Cmd.cmd_func()` to `Cmd.get_command_func()`.
+    - `cmd2` no longer sets a default title for a subparsers group. If you desire a title, you will
+      need to pass one in like this `parser.add_subparsers(title="subcommands")`. This is standard
+      `argparse` behavior.
 - Enhancements
     - New `cmd2.Cmd` parameters
         - **auto_suggest**: (boolean) if `True`, provide fish shell style auto-suggestions. These